Completed palindrome function
I filled in the palindrome function and I also added the onlyLetters function from the codingbat problem set. I added some spacing for the reverse function so that it's much more clean and you can read it better. There is also spacing around each function for the same purpose of readability.  The program now works for the text set and should work for other palindromes as well but I'm not too sure.